requirements.txt
# Core dependencies
streamlit>=1.28.0
pillow>=10.0.0
numpy>=1.23.0
pyyaml>=6.0

# OpenCV (headless version for cloud deployment)
opencv-python-headless>=4.8.0

# YOLO and dependencies
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# OCR dependencies
tensorflow>=2.10.0
